Task #4805: Merged latest trunk into the branch and resolved merge conflict in DelayAndBandPassKernel.cc

+++ branches/Cobalt-Task4805/RTCP/Cobalt/GPUProc/src/cuda/Kernels/BeamFormerKernel.h
@@ -74,7 +74,7 @@ namespace LOFAR
                              const Parameters &param);
 
       void enqueue(const BlockID &blockId, PerformanceCounter &counter,
-                   float subbandFrequency, unsigned SAP);
+                   double subbandFrequency, unsigned SAP);
 
     private:
       // Dump output buffers of a given kernel to disk. Use \a blockId to
